requirements.txt
dj-database-url==0.4.0
Django==1.9.2
django-registration-redux==1.4
gunicorn==19.4.5
psycopg2==2.6.1
whitenoise==2.0.6
ipython==4.1.2
ipython-genutils==0.1.0
Logbook==0.12.5
nose==1.3.7
Quandl==2.8.9
numpy==1.10.4
pandas==0.17.1
pandas_datareader==0.2.1
python-dateutil==2.4.2
pytz==2015.7
pyzmq==15.2.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.8.0
requests==2.4.3
ystockquote==0.2.4
django-crispy-forms==1.6.0
django-nose==1.4.3
coverage==3.7.1
mock==1.3.0
feedparser==5.2.1
selenium==2.53.1
sendgrid==2.0.0
isort==4.2.5
flake8==2.5.4
sphinx==1.4.1
coveralls==1.1
sphinx_rtd_theme==0.1.9
